+++ src/utils/api.ts
@@ -3,9 +3,10 @@ import {
   useMutation,
   UseQueryOptions,
   UseMutationOptions,
-  useQueryClient,
 } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
+import { useToken } from '@/hooks/useToken';
+
 export const API_DOMAIN = import.meta.env.VITE_API_DOMAIN;
 type RequestMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
 
@@ -142,26 +143,4 @@ export const usePatchRequest = <T, Variables>(
     },
     ...options,
   });
-};
-
-export const useToken = () => {
-  const queryClient = useQueryClient();
-
-  const accessToken =
-    queryClient.getQueryData<string>(['accessToken']) ||
-    localStorage.getItem('accessToken');
-
-  const setAccessToken = (token: string) => {
-    queryClient.setQueryData(['accessToken'], token);
-    localStorage.setItem('accessToken', token);
-  };
-
-  const clearTokens = () => {
-    queryClient.removeQueries({ queryKey: ['accessToken'] });
-    queryClient.removeQueries({ queryKey: ['refreshToken'] });
-    localStorage.removeItem('accessToken');
-    localStorage.removeItem('refreshToken');
-  };
-
-  return { accessToken, setAccessToken, clearTokens };
 };
